fix: move useVideoTexture to /web
fixes #1917.
Moves useVideoTexture to /web. Also fix reference from faceControls.
